ppc4xx: Update Makalu defconfig to use device-tree booting as default
This patch reworks the default environment on Makalu. Now "net_nfs" for
example uses the device-tree style booting formerly know as "net_nfs_fdt".
Also the addresses in RAM were changed because of the new image booting
support, which check for image overwriting. So the addresses needed to
get adjusted.

Signed-off-by: Stefan Roese <sr@denx.de>